feat(FE): #29 recoil 상태관리 적용해 코드 일부 개선 (테이블, 모달, 버튼에 사용되는 checkedList state 관리위해 recoil state 분리), apollo client (graphql 포함, 응답 테스트 완료) 설치 및 관련 설정 일부 변경, 중복 type interface 분리
