+++ frontend/.eslintrc.js
@@ -50,6 +50,14 @@ module.exports = {
     'no-else-return': ['error', { allowElseIf: true }],
     'operator-linebreak': 'off',
     'linebreak-style': 'off',
+    'react-hooks/rules-of-hooks': 'error',
+    'react-hooks/exhaustive-deps': [
+      'warn',
+      {
+        additionalHooks: 'useRecoilCallback',
+      },
+    ],
+    'import/prefer-default-export': 'off',
   },
   settings: {
     'import/resolver': {
