+++ frontend/package.json
@@ -3,21 +3,24 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@apollo/client": "^3.6.5",
     "@chakra-ui/react": "^1.8.8",
     "@testing-library/jest-dom": "^5.16.4",
     "@testing-library/react": "^12.1.5",
     "@testing-library/user-event": "^13.5.0",
-    "@toast-ui/react-editor": "^3.1.7",
     "@toast-ui/editor": "^3.1.7",
+    "@toast-ui/react-editor": "^3.1.7",
     "@types/jest": "^27.5.1",
     "@types/node": "^16.11.34",
     "@types/react": "^18.0.9",
     "@types/react-dom": "^18.0.4",
+    "graphql": "^16.5.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-icons": "^4.3.1",
     "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.1",
+    "recoil": "^0.7.3-alpha.2",
     "typescript": "^4.6.4",
     "web-vitals": "^2.1.4"
   },
