+++ frontend/src/recoil/index.ts
@@ -1,0 +1,6 @@
+import { atom } from 'recoil';
+
+export const checkedListState = atom<number[]>({
+  key: 'checkedListState',
+  default: [],
+});
