+++ frontend/src/types/index.ts
@@ -1,0 +1,19 @@
+interface ListItem {
+  id: number;
+  subject: string;
+  status: string;
+  register: string;
+  createdAt: string;
+}
+
+interface Page {
+  page: number;
+  perPage: number;
+  totalCount: number;
+  totalPage: number;
+}
+
+export type {
+  ListItem,
+  Page
+};
